requirements.txt
cryptography>=3.4.8
numpy>=1.21.0
argon2-cffi>=21.3.0